colouring: fixing type that was fixed before (wait_counter cannot be usigned in current impl)

+++ framework/src/main/algorithms/Colouring.h
@@ -37,7 +37,7 @@ namespace details {
 		VertexTempData() : wait_counter(0) {}
 
 		std::set<VertexColour> used_colours;
-		size_t wait_counter;
+		long long int wait_counter;
 	};
 }
